renaming range histogram, adding options for histogram bounds in tools command, adding enumerate tools command
