+++ docs/user/commandline_tools.rst
@@ -32,9 +32,10 @@ Run ``geomesa`` without any arguments to produce the following usage text::
         list                List GeoMesa features for a given catalog
         queryrasterstats    Export queries and statistics about the last X number of queries to a CSV file.
         stats-analyze       Analyze statistics on a GeoMesa feature type
-        stats-bounds        View bounds on attributes in a GeoMesa schema
-        stats-count         View feature counts in a GeoMesa schema
-        stats-histogram     View statistics on a GeoMesa feature type
+        stats-bounds        View or calculate bounds on attributes in a GeoMesa feature type
+        stats-count         Estimate or calculate feature counts in a GeoMesa feature type
+        stats-enumerate     Enumerate attribute values in a GeoMesa feature type
+        stats-histogram     View or calculate counts of attribute in a GeoMesa feature type, grouped by sorted values
         removeschema        Remove a schema and associated features from a GeoMesa catalog
         tableconf           Perform table configuration operations
         version             GeoMesa Version
@@ -492,6 +493,25 @@ Example usage::
       Running stat query...
       Count: 182
 
+
+stats-enumerate
+~~~~~~~~~~~~~~~
+
+Enumerates the values for attributes in your data set. You can enumerate all values, or only values for
+features that match a CQL filter.
+
+Example usage::
+
+    $ geomesa stats-enumerate -u username -p password -i instance -z zoo1,zoo2,zoo3 \
+        -c geomesa.data -f twitter -a user_id
+      Running stat query...
+      Values for 'user_id':
+        3144822634 (26383)
+        388009236 (20457)
+        497145453 (19514)
+        563319506 (15848)
+        2841269945 (15716)
+        ...
 
 stats-histogram
 ~~~~~~~~~~~~~~~
